moved confetti to lib folder
