+++ lib/confetti.js
